requirements.txt
streamlit==1.49.0
tensorflow==2.19.0
numpy==2.0.2
pillow==11.3.0
gdown